Add files via upload
Particle Catalogue Project Report

This report provides an in-depth analysis of the Particle Catalogue Project, a C++ simulation that models Standard Model particles. Written by Luca Vicaria for the University of Manchester's Department of Physics and Astronomy, this report details the use of object-oriented programming to encapsulate and simulate the properties and behaviors of fundamental particles. The project implements advanced C++ features such as templates, exception handling, and the Standard Template Library (STL) to create a dynamic and robust particle catalogue.

The report includes a comprehensive overview of the class hierarchy, which organizes particles into quarks, leptons, and bosons, and it delves into the use of smart pointers and custom containers for efficient memory management. It also addresses challenge features, including a physically consistent implementation of the four-momentum vector, which ensures accurate invariant mass calculations. The project demonstrates error-handling mechanisms, lambda functions, and efficient data management techniques, illustrating how these C++ features contribute to a scalable and maintainable codebase.

This report serves as a detailed guide for understanding the project's design, code structure, and functionality, making it a valuable resource for both educational and research purposes in the field of particle physics.